% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTAIRid.R
\name{f_allid}
\alias{f_allid}
\title{get_all_id's function}
\usage{
f_allid(querytext)
}
\arguments{
\item{querytext}{A string containing one or more TAIR ids}
}
\description{
Extracts all TAIR'ids from any texts. Does not work well when applied to columns
}
