% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goenrich_nonmodel.R
\name{f_gene2go}
\alias{f_gene2go}
\title{gene2go}
\usage{
f_gene2go(df_gene2go)
}
\arguments{
\item{df_gene2go}{takes a dataframe of two columns. The first column should contain only single geneids. The second column should contain a string containing GO:terms separated by ",".}
}
\description{
Converts a two column dataframe to a named list (as required for goseq_nm).
}
