% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goenrich_nonmodel.R
\name{f_genelength}
\alias{f_genelength}
\title{Extract gene length from fasta file}
\usage{
f_genelength(fasta_file)
}
\arguments{
\item{fasta_file}{Path to a fasta file containing the genes of interest}
}
\description{
Extracts length of genes from fasta file into a named list.
}
\details{
uses biostrings to import fastafiles
}
