% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goenrich_nonmodel.R
\name{f_go2geneid}
\alias{f_go2geneid}
\title{go2geneid}
\usage{
f_go2geneid(geneid2go)
}
\arguments{
\item{geneid2go}{takes a named list of GO-terms. List elements should be GO-terms. Names are corresponding geneids.}
}
\description{
generates the inverse of geneid2go
}
\details{
takes a named list of GO-terms, as generated by geneid2go. Generates the inverse of the input.
}
