% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTAIRid.R
\name{f_singleid}
\alias{f_singleid}
\title{get single tair id}
\usage{
f_singleid(querytext)
}
\arguments{
\item{querytext}{A string containing one tair id}
}
\description{
extracts single tair id; since it does not generate a list this funciton gets along better with dataframe operations.
}
