% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPEAKS.R
\name{get_peaks}
\alias{get_peaks}
\title{get_peaks}
\usage{
get_peaks(fl_folder, chromosome, start_position_gene, length_upstream)
}
\arguments{
\item{fl_folder}{directory containining the .NarrowPeak files. The function runs recursive by default (non-recursive currently not supported). When using this function on O'Malley's dataset set the path to "dap_download_may2016_peaks".}

\item{chromosome}{chromosome number of the section of interest.}

\item{start_position_gene}{The position on the chromosome used as starting position of the section of interest. If only interested in the promotor use startcodon. If also interested in potenial binding sites in the introns/coding regions provide the position of the stopcodong or 3' UTR.}

\item{length_upstream}{The length of promotor sequeance of interest. Provide a negative number when querying for a gene in reverse orientation.}
}
\value{
dataframe containing peaks and associated transcriptionfactors
}
\description{
Loops through .NarrowPeaks files in input folder and extracts the peaks in the specified (promotor) secition of the genome.
}
\details{
This function has been independently developed to query files from the O'Malley paper ( doi: 10.1016/j.cell.2016.04.038).
Download input data from http://neomorph.salk.edu/dap_web/pages/browse_table_aj.php; in .NarrowPeaks format
}
\author{
Kilian Duijts
}
