% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPEAKS.R
\name{get_tf_from_peaks}
\alias{get_tf_from_peaks}
\title{get_tf_from_peaks}
\usage{
get_tf_from_peaks(fl_folder, chromosome, start_position_gene, length_upstream)
}
\arguments{
\item{fl_folder}{directory containining the .NarrowPeak files. The function runs recursive by default (non-recursive currently not supported). When using this function on O'Malley's dataset set the path to "dap_download_may2016_peaks".}

\item{chromosome}{chromosome number of the section of interest.}

\item{start_position_gene}{The position on the chromosome used as starting position of the section of interest. If only interested in the promotor use startcodon. If also interested in potenial binding sites in the introns/coding regions provide the position of the stopcodong or 3' UTR.}

\item{length_upstream}{The length of promotor sequeance of interest. Provide a negative number when querying for a gene in reverse orientation.

@return dataframe containing associated transcription factor geneid's and symbols.

@author Kilian Duijts}
}
\description{
extracts only the unique geneids and symbols of transcription factors binding in the specified section.
}
\details{
This function uses get_peaks but returns a dataframe with only the unique geneids and names for the peaks in the section of interest.

parameters are identical to get_peaks.
}
