% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTAIRid.R
\name{gettairid_all}
\alias{gettairid_all}
\title{Get TAIR identifier}
\usage{
gettairid_all(querytext, exclude_plastid, exclude_nuclear, specify)
}
\arguments{
\item{querytext}{The text (character string) containing the gene identifiers. Make sure the character string does not contain quotation marks within.}

\item{exclude_plastid}{OPTIONAL - When TRUE excludes identifiers of plastid genes from output (mitochondrial or chloroplastic).}

\item{exclude_nuclear}{OPTIONAL - When TRUE excludes identifiers of nuclear genes from output.}

\item{specify}{OPTIONAL - Allows you to specify chromosome numbers or "M" or "C" to be included in output.
Takes character vector of chromosome numbers and/or "M" and "C" for mitochondrial or chloroplastic gene identifiers respectively.
remember to put chromosome numbers in quotation marks.


When OPTIONAL variables are left empty all gene identifiers are extracted.}
}
\value{
A list of TAIR identifiers found in the querytext.
}
\description{
Extract arabidopsis (TAIR) gene identifiers from a text.
}
\details{
These identifiers conform to the following structure:

"AT" followed by the 

Chromosome number, a "M" for mitochondria or a "C" for chloroplast.

"G" followed by a
five-digit code, numbered from top/north to bottom/south of chromosome

Example of gene id: 
nuclear: AT3G28291
mitochondrial: ATMG00730

Does not extract: transposable elements (these do not follow this naming convention)
For more info on TAIR ID nomenclature see https://www.arabidopsis.org/portals/nomenclature/guidelines.jsp
}
\examples{
q<-"AT1G19030 AT3G61800,AT2G02200convallisAT5G23670,ATMG00730loreAT3G28291,AT4G07460,AT3G29ATMG00730"

ids<-gettairid(querytext=q)
ids:
"AT1G19030" "AT3G61800" "AT2G02200" "AT5G23670" "ATMG00730" "AT3G28291" "AT4G07460" "ATMG00730"

ids<-gettairid(querytext = q, exclude_nuclear = TRUE)
ids: "ATMG00730" "ATMG00730"

ids<-gettairid(querytext = q, specify = c("3", "5"))
ids: "AT3G61800" "AT5G23670" "AT3G28291"



}
\author{
Kilian Duijts
}
