% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goenrich_nonmodel.R
\name{non_model_goseq}
\alias{non_model_goseq}
\title{Gene Ontology testing for non-model species}
\usage{
non_model_goseq(querylist, padj, geneid2GO, cdna)
}
\arguments{
\item{querylist}{A list with for each element a non-model species gene id.}

\item{padj}{Numerical value specifying the cut-off for the adjusted pvalue

# method <currently not available>}

\item{geneid2GO}{A named list for which the names are geneids and the list elements consist of GO-terms (like so: GO:0006355). Please make sure your geneid's are identical to the query genes.}

\item{cDNA}{fasta file containing the non-model species geneid's and their sequences. The gene lenght is used for the probabily weight; normalization for gene length.}
}
\description{
Re-application of the goseq ("https://bioconductor.org/packages/release/bioc/html/goseq.html") package for use on non model species.
This is based on a tutorial ("https://knozue.github.io/post/2018/10/17/over-representation-analysis-3-goseq-with-non-model-go-term.html") by knozue
}
\details{
The function makes it easier to test gene sets of non-model species.

It takes the following parameters:
}
\author{
Kilian Duijts
}
