% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gettairid.R
\name{gettairid}
\alias{gettairid}
\title{Get TAIR identifier}
\usage{
gettairid(querytext)
}
\arguments{
\item{querytext}{The text (character string) containing the gene identifiers. Make sure it does not contain any quotationmarks ("") except at the start and end of the string.}
}
\value{
A list of TAIR identifiers found in the querytext.
}
\description{
Extract arabidopsis (TAIR) gene identifiers from a text. These identifiers conform to the following structure:
}
\details{
"AT" 
followed by the chromosome number the gene is located on; 

or alternatively a M for mitochondria or C for chloroplast if the gene is located on one of the plastid genomes.
}
\examples{
ids<-gettairid(querytext="AT1G19030 AT3G61800,AT2G02200convallisAT5G23670,ATMG00730loreAT3G28291,AT4G07460,AT3G29ATMG00730"




}
