% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaRdenbox.R
\name{gettairid}
\alias{gettairid}
\title{Get tair identifier}
\usage{
gettairid(querytext)
}
\arguments{
\item{querytext}{The text containing the gene identifiers.}
}
\value{
A list of Tair identifiers found in the querytext
}
\description{
Extract arabidopsis (TAIR) gene identifiers from a text. These identifiers conform to the following structure:
AT followed by the chromosome number the gene is located on; 
or alternatively a M for mitochondria or C for chloroplast if the gene is located on one of the plastid genomes.
}
\examples{
ids<-gettairid(querytext=c("AT1G19030 AT3G61800,AT2G02200convallisAT5G23670,ATMG00730loreAT3G28291,AT4G07460,AT3G29ATMG00730"))




}
